Implement JSON-RPC
Add lsp_endpoint, a JSON-RPC parser and serailizer with the interactions
with other classes:

* Dispatch incoming requests and notifications to an lsp_endpoint_handler
* Send outgoing requests and notifications (given by a
  lsp_endpoint_handler) to an lsp_endpoint_remote (e.g. an
  lsp_pipe_writer)

lsp_endpoint will be used for quick-lint-js' Language Server Protocol
implementation, hence 'lsp' in the class' name.

lsp_endpoint is currently only used in tests. It is not integrated with
the main quick-lint-js executable yet. In particular, an
lsp_endpoint_handler class needs to be implemented before lsp_endpoint
is useful.